+++ Project FG/Assets/0_ProjectFG/Scripts/Interface/IDamageable.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public interface IDamageable 
 {
-    public int Health { get; }
+    public float Health { get; }
     public bool IsDie {  get; }
-    public void OnDamage(int damage);
+    public void OnDamage(float damage);
 
     public void Die();
 }
